Rename file to in_diff.rs
