+++ src/main.rs
@@ -6,9 +6,9 @@ mod build_dir;
 mod cargo;
 mod config;
 mod console;
-mod diff_filter;
 mod exit_code;
 mod fnvalue;
+mod in_diff;
 mod interrupt;
 mod lab;
 mod list;
@@ -45,7 +45,7 @@ use tracing::debug;
 
 use crate::build_dir::BuildDir;
 use crate::console::Console;
-use crate::diff_filter::diff_filter;
+use crate::in_diff::diff_filter;
 use crate::interrupt::check_interrupted;
 use crate::lab::test_mutants;
 use crate::list::{list_files, list_mutants, FmtToIoWrite};
